Add callback to be called after `set()`s
Add a callback argument to the constructor that will be called after a
`set()` call results in a new object. The callback is called with the
new object as its argument. The callback is stored on the object as an
inenumerable property, `__callback`. This causes multiple calls to
`immutable()` to not return the strictly equal objects. Tests expected
this behavior were changed.

Change the constructor returned by `createClass()` to always use the
`ImmutableObject` constructor, passing the last argument as the
callback.